requirements.txt
python-telegram-bot==8.1.1
pymongo==3.6.0
tensorflow==1.3.0
Cython==0.27.2
Pillow==4.3.0